move to desired feature folder
